C program that prints the size of various types on the computer it is compiled and run on.

+++ 0x00-hello_world/6-size.c
@@ -13,10 +13,10 @@ int main(void)
 	long long int d;
 	float f;
 
-	printf("Size of char:%lu byte(s)\n",(unsigned long)sizeof(a));
-	printf("Size of int:%lu byte(s)\n",(unsigned long)sizeof(b));
-	printf("Size of long int:%lu byte(s)\n",(unsigned long)sizeof(c));
-	printf("Size of long long int:%lu byte(s)\n",(unsigned long)sizeof(d));
-	printf("Size of float:%lu byte(s)\n",(unsigned long)sizeof(f));
+	printf("Size of char:%d byte(s)\n",sizeof(a));
+	printf("Size of int:%d byte(s)\n",sizeof(b));
+	printf("Size of long int:%d byte(s)\n",sizeof(c));
+	printf("Size of long long int:%d byte(s)\n",sizeof(d));
+	printf("Size of float:%d byte(s)\n",sizeof(f));
 	return (0);
 }
